# Unify execution state into a single struct