--- a/src/wasm-lib/tests/executor/inputs/no_visuals/pipe_substitution_inside_function_called_from_pipeline.kcl
+++ b/src/wasm-lib/tests/executor/inputs/no_visuals/pipe_substitution_inside_function_called_from_pipeline.kcl
@@ -1,0 +1,7 @@
+// Make sure pipe value doesn't leak into the function call.
+fn f = (ignored) => {
+  return %
+}
+
+const answer = %
+  |> f(%)