This combines `ProgramMemory`, `DynamicState`, and `PipeInfo` into a single struct `ExecState`. This greatly simplifies how we thread state through the Executor since it all happens on a single struct. This also gives us a single place to add fields when we want to add more state. (Wanting to add more state was what triggered this change.)

Because we use `&mut ExecState`, we're able to greatly reduce the amount of cloning we do. And when we still clone, it's much clearer _why_ we need to clone.

I didn't attempt to combine `ExecutorContext` since I already changed so much, but in theory, we should be able to combine that also.

Related to #3238. (Doesn't close it.)

Checklist:

- [x] Test pipe value is not available inside the function of a function call
- [x] Benchmark to make sure performance hasn't regressed. Running `executor_benchmark_criterion` locally indicates no significant change in performance. I've observed it running both slightly faster and slightly slower than `main`. Is variance due to the engine?
- [x] Merge `main`. Needs updates in `array.rs`.

Chalmers benchmarked it using #3898.

```
this branch:
355.0 ms ±   3.2 ms
main:
344.9 ms ±   2.3 ms
```