@adamchalmers 2024-09-16
@jtran fyi most reliable way to benchmark would be a test in `no_visuals.rs` -- they don't start an engine connection. You can then use `hyperfine` to time `cargo nextest run -- test_filter` repeatedly.